requirements.txt
watchdog==6.0.0
pyyaml